# docs/reference/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_api_ingest_v1_ingest_post": {
        "properties": {
          "embed_model": {
            "default": "bge-small-en-v1.5",
            "title": "Embed Model",
            "type": "string"
          },
          "index_out": {
            "title": "Index Out",
            "type": "string"
          },
          "input_path": {
            "title": "Input Path",
            "type": "string"
          }
        },
        "required": [
          "input_path",
          "index_out"
        ],
        "title": "Body_api_ingest_v1_ingest_post",
        "type": "object"
      },
      "Body_api_query_v1_query_post": {
        "properties": {
          "bundle_k": {
            "default": 5,
            "title": "Bundle K",
            "type": "integer"
          },
          "embed_model": {
            "default": "bge-small-en-v1.5",
            "title": "Embed Model",
            "type": "string"
          },
          "index_path": {
            "title": "Index Path",
            "type": "string"
          },
          "kneighbors": {
            "default": 6,
            "title": "Kneighbors",
            "type": "integer"
          },
          "lamC": {
            "default": 0.5,
            "title": "Lamc",
            "type": "number"
          },
          "lamP": {
            "default": 0.0,
            "title": "Lamp",
            "type": "number"
          },
          "lamQ": {
            "default": 4.0,
            "title": "Lamq",
            "type": "number"
          },
          "q": {
            "title": "Q",
            "type": "string"
          },
          "tol": {
            "default": 0.001,
            "title": "Tol",
            "type": "number"
          }
        },
        "required": [
          "index_path",
          "q"
        ],
        "title": "Body_api_query_v1_query_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Params": {
        "properties": {
          "deterministic_k": {
            "default": false,
            "title": "Deterministic K",
            "type": "boolean"
          },
          "kneighbors": {
            "default": 6,
            "title": "Kneighbors",
            "type": "integer"
          },
          "lamC": {
            "default": 0.5,
            "title": "Lamc",
            "type": "number"
          },
          "lamG": {
            "default": 1.0,
            "title": "Lamg",
            "type": "number"
          },
          "lamP": {
            "default": 0.0,
            "title": "Lamp",
            "type": "number"
          },
          "lamQ": {
            "default": 4.0,
            "title": "Lamq",
            "type": "number"
          },
          "neighbor_seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neighbor Seed"
          }
        },
        "title": "Params",
        "type": "object"
      },
...
      "SettleOptions": {
        "properties": {
          "bundle_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle K"
          },
          "dt": {
            "default": 1.0,
            "title": "Dt",
            "type": "number"
          },
          "include_receipt": {
            "default": true,
            "title": "Include Receipt",
            "type": "boolean"
          },
          "max_iters": {
            "default": 12,
            "title": "Max Iters",
            "type": "integer"
          },
          "tol": {
            "default": 0.001,
            "title": "Tol",
            "type": "number"
          }
        },
        "title": "SettleOptions",
        "type": "object"
      },
      "SettleRequest": {
        "properties": {
          "Y": {
            "description": "Matrix N x D (list of rows)",
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Y",
            "type": "array"
          },
          "chain": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain"
          },
          "gates": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gates"
          },
          "options": {
            "$ref": "#/components/schemas/SettleOptions",
            "default": {
              "dt": 1.0,
              "include_receipt": true,
              "max_iters": 12,
              "tol": 0.001
            }
          },
          "params": {
            "$ref": "#/components/schemas/Params",
            "default": {
              "deterministic_k": false,
              "kneighbors": 6,
              "lamC": 0.5,
              "lamG": 1.0,
              "lamP": 0.0,
              "lamQ": 4.0
            }
          },
          "psi": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Psi"
          }
        },
        "required": [
          "Y"
        ],
        "title": "SettleRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Oscillink Cloud API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/bundle": {
      "post": {
        "description": "Return only the bundle (requires options.bundle_k).",
        "operationId": "bundle_v1_bundle_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bundle"
      }
    },
...
    "/v1/ingest": {
      "post": {
        "operationId": "api_ingest_v1_ingest_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-License-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-License-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_ingest_v1_ingest_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Ingest V1 Ingest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Ingest",
        "tags": [
          "ingest"
        ]
      }
    },
...
    "/v1/query": {
      "post": {
        "operationId": "api_query_v1_query_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-License-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-License-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_query_v1_query_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Query V1 Query Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Query",
        "tags": [
          "ingest"
        ]
      }
    },
    "/v1/receipt": {
      "post": {
        "description": "Return only the receipt (always include_receipt).",
        "operationId": "receipt_v1_receipt_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receipt"
      }
    },
...
  }
}